Appropriate error code is returned if consumer or producer of an IPC channel is not initialized.
Error code -ERR_CONSUMER_NOT_INIT is returned if consumer of the IPC channel is
not initialized and producer is trying to produce data on that channel.
Similarly, error code -ERR_PRODUCER_NOT_INIT is returned if producer of the
channel is not initialized and consumer is trying to consume data from that
channel.

Signed-off-by: Naveen Burmi <NaveenBurmi@freescale.com>