Multi RAT support in IPC Library
This Patch adds the support of multiple RATs in the IPC library.
For each rat_id there would exist a separate ipc instance with its own set of channels.

The only change from user api is the fsl_ipc_init which now takes a rat_id as a input parameter.

Signed-off-By: Manish Jaggi <manish.jaggi@freescale.com>